Add support for AxisBase, ValueAxis and CategoryAxis for API, JSO, JSNI and JSON Clean up code for demo into sub-packages Code generation SubClassJSNIWriter to provide jsni code generation for classes like ValueAxis and CategoryAxis which extend from AxisBase

+++ src/com/amcharts/impl/AmChart.java
@@ -335,12 +335,12 @@ public class AmChart extends Composite implements IsAmChart, IJavaScriptWrapper<
 	}-*/;
 
 	@Override
-	public native Integer getHideBalloonTime() /*-{
+	public native double getHideBalloonTime() /*-{
 		return this.@com.amcharts.impl.AmChart::jso.hideBalloonTime;
 	}-*/;
 
 	@Override
-	public native void setHideBalloonTime( Integer hideBalloonTime ) /*-{
+	public native void setHideBalloonTime( double hideBalloonTime ) /*-{
 		this.@com.amcharts.impl.AmChart::jso.hideBalloonTime = hideBalloonTime;
 	}-*/;
